Nitrogen:  Pass unsigned long* to Delay() for LP64

+++ nitric/Nitrogen/OSUtils.hh
@@ -137,9 +137,10 @@ namespace Nitrogen
 	
 	// SwapMMUMode
 	
-	inline UInt32 Delay( UInt32 ticks )
+	inline
+	unsigned long Delay( unsigned long ticks )
 	{
-		UInt32 finalTicks;
+		unsigned long finalTicks;
 		::Delay( ticks, &finalTicks );
 		
 		return finalTicks;
